Rearrange and clean up of CTRNN and Decide code.

+++ src/main/java/net/happybrackets/patternspace/ctrnn/Chromosome.java
@@ -14,11 +14,11 @@ public class Chromosome implements Serializable {
 	
 	private static final long serialVersionUID = 1;
 	
-	private float[] genes;
+	private double[] genes;
 	private int index;
 	
 	public Chromosome(int length) {
-		genes = new float[length];
+		genes = new double[length];
 		reset();
 	}
 	
@@ -26,7 +26,7 @@ public class Chromosome implements Serializable {
 		index = 0;
 	}
 	
-	public float getNextGene() {
+	public double getNextGene() {
 		return genes[index++];
 	}
 	
@@ -34,7 +34,7 @@ public class Chromosome implements Serializable {
 		Gene[] sourceGenes = source.getGenes();
 		Chromosome c = new Chromosome(sourceGenes.length);
 		for(int i = 0; i < sourceGenes.length; i++) {
-			c.genes[i] = (float)((DoubleGene)sourceGenes[i]).doubleValue();
+			c.genes[i] = ((DoubleGene)sourceGenes[i]).doubleValue();
 		}
 		return c;
 	}
@@ -49,31 +49,31 @@ public class Chromosome implements Serializable {
 	public static Chromosome newRandom(Random rng, int length) {
 		Chromosome newRandom = new Chromosome(length);
 		for(int i = 0; i < length; i++) {
-			newRandom.genes[i] = rng.nextFloat();
+			newRandom.genes[i] = rng.nextDouble();
 		}
 		return newRandom;
 	}
 	
-	public float[] getGenes() {
+	public double[] getGenes() {
 		return genes;
 	}
 	
-	public Chromosome mutate(Random rng, float stdDev) {
+	public Chromosome mutate(Random rng, double stdDev) {
 		Chromosome child = new Chromosome(genes.length);
 		for(int i = 0; i < genes.length; i++) {
-			child.genes[i] = genes[i] + (float)rng.nextGaussian() * stdDev;
+			child.genes[i] = genes[i] + rng.nextGaussian() * stdDev;
 			if(child.genes[i] < 0.0f) child.genes[i] = -0.5f * child.genes[i];
 			else if(child.genes[i] > 1.0f) child.genes[i] = 1.0f - 0.5f * (child.genes[i] - 1.0f);
 		}
 		return child;
 	}
 
-	public float getGeneticDistance(Chromosome c) {
-		float distance = 0.0f;
+	public double getGeneticDistance(Chromosome c) {
+		double distance = 0.0f;
 		for(int i = 0; i < genes.length; i++) {
 			distance += (genes[i] - c.genes[i]) * (genes[i] - c.genes[i]);
 		}
-		distance = (float)Math.sqrt(distance / genes.length); //RMS
+		distance = Math.sqrt(distance / genes.length); //RMS
 		return distance;
 	}
